pyproject.toml
[build-system]
build-backend = "uv_build"
requires = ["uv_build>=0.9.26"]

[dependency-groups]
dev = [
  'calcipy[ddict,doc,lint,nox,tags,test,types]>=5.0.0',
  # 'calcipy[dev]>=6.0.0',  # PLANNED: When available
  'pytest-asyncio>=0.24.0',
  'pytest-structlog>=1.1',
  'pytest-subprocess>=1.5.2',
  'structlog>=24.4.0',
  'types-psutil>=6.0.0.20240901',
  'types-pyyaml>=6.0.12.20240917',
]

[project]
authors = [
  {email = "dev.act.kyle@gmail.com", name = "Kyle King"},
]
classifiers = [
  "Development Status :: 1 - Planning",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
] # https://pypi.org/classifiers/
dependencies = [
  "tomli>=2.0.2; python_version < '3.11'",
  'beartype>=0.19.0',
  'corallium>=2.0.1',
  'rich>=13.9.2',
  'typing-extensions>=4.12.2',
]
description = "Shared functionality for the calcipy-ecosystem"
keywords = ["calcipy_template"]
license = "MIT"
maintainers = []
name = "corallium"
readme = "docs/README.md"
requires-python = ">=3.10.11"
version = "2.1.1"

[project.urls]
"Bug Tracker" = "https://github.com/kyleking/corallium/issues"
"Changelog" = "https://github.com/kyleking/corallium/blob/main/docs/docs/CHANGELOG.md"
"Documentation" = "https://corallium.kyleking.me"
"Repository" = "https://github.com/kyleking/corallium"

[tool.commitizen]
version = "2.1.1"
version_files = ["corallium/__init__.py:^__version", "pyproject.toml:^version"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
enable_error_code = ["ignore-without-code", "possibly-undefined", "redundant-expr", "truthy-bool"]
extra_checks = true
files = ["corallium", "tests"]
no_implicit_reexport = true
plugins = [
]
python_version = "3.10"
show_column_numbers = true
show_error_codes = true
strict_equality = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.poetry.extras]
tags = ["arrow"]

[tool.pyright]
include = ["corallium", "tests"]
pythonVersion = "3.10"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
# Docs: https://github.com/charliermarsh/ruff
# Tip: uv run ruff rule RUF100
line-length = 120
target-version = 'py310'

[tool.ruff.format]
quote-style = 'single'

[tool.ruff.lint]
ignore = [
  'ANN002', # Missing type annotation for `*args`
  'ANN003', # Missing type annotation for `**kwargs`
  'ANN101', # Missing type annotation for `self` in method (automatically inferred)
  'ANN102', # Missing type annotation for `cls` in classmethod (automatically inferred)
  'ANN401', # Dynamically typed expressions (typing.Any) are disallowed in `pop_key`
  'BLE001', # Do not catch blind exception: `Exception`
  'COM812', # Conflicts with formatter
  'CPY001', # Missing copyright notice at top of file
  'D203', # "1 blank line required before class docstring" (Conflicts with D211)
  'D213', # "Multi-line docstring summary should start at the second line" (Conflicts with D212)
  'DOC501', # Raised exception `RuntimeError` missing from docstring
  'EM101', # Exception must not use a string literal, assign to variable first
  'FIX001', # Line contains FIXME
  'FIX002', # Line contains TODO
  'FIX004', # Line contains HACK
  'PLR0913', # Too many arguments in function definition (6 > 5)
  'S404', # `subprocess` module is possibly insecure
  'S602', # `subprocess` call with `shell=True` identified, security issue
  'S604', # Function call with `shell=True` parameter identified, security issue
  'TCH001', # Move application import into a type-checking block (Conflicts with Beartype)
  'TCH002', # Move third-party import into a type-checking block (Conflicts with Beartype)
  'TCH003', # Move standard library import into a type-checking block (Conflicts with Beartype)
  'TD001', # Invalid TODO tag: `FIXME`
  'TD002', # Missing author in TODO; try: `# TODO(<author_name>): ...`
  'TD003', # Missing issue link on the line following this TODO
  'TRY003', # Avoid specifying long messages outside the exception class
]
preview = true
select = ['ALL']
unfixable = [
  'ERA001', # Commented out code
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = 'single'

[tool.ruff.lint.flake8-tidy-imports.banned-api]
'invoke.collection.Collection'.msg = 'Use calcipy.collection.Collection instead.'
'invoke.tasks.task'.msg = 'Use calcipy.cli.task instead.'
'typing.Callable'.msg = 'Use "from collections.abc" instead.'
'typing.Dict'.msg = 'Use "from __future__ import annotations"'
'typing.List'.msg = 'Use "from __future__ import annotations"'
'typing.Optional'.msg = 'Use "from __future__ import annotations"'
'typing.Protocol'.msg = 'Use "from beartype.typing" instead.'

[tool.ruff.lint.isort]
known-first-party = ['corallium']

[tool.ruff.lint.per-file-ignores]
'./corallium/../*.py' = [
  'INP001', # File `/<>.py` is part of an implicit namespace package. Add an `__init__.py`.
]
'__init__.py' = [
  'D104', # Missing docstring in public package
]
'scripts/*.py' = [
  'INP001', # File `scripts/*.py` is part of an implicit namespace package. Add an `__init__.py`.
]
'scripts/check_imports.py' = [
  'F401', # imported but unused; consider adding to __all__ or using a redundant alias
]
'tests/*.py' = [
  'ANN001', # Missing type annotation for function argument
  'ANN201', # Missing return type annotation for public function
  'ANN202', # Missing return type annotation for private function `test_make_diffable`
  'ARG001', # Unused function argument: `line`
  'D100', # Missing docstring in public module
  'D103', # Missing docstring in public function
  'PLC2701', # Private name import `_<>` from external module
  'PT004', # flake8-pytest-style: fixture does not return
  'S101', # Use of `assert` detected
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.tomlsort]
all = true
in_place = true
trailing_comma_inline_array = true

[tool.uv]
default-groups = ["dev"]
required-version = ">=0.9.0"

[tool.uv.build-backend]
module-name = "corallium"
module-root = ""
